refactor(web): make GitHub org membership a single source of truth
Collapse the sign-in policy so GitHub org membership has one owner.

- Delete the dead synchronous `activeOrganizations` branch from
  getAccessAllowReason: the sign-in callback never fed it, so the async
  checkGitHubOrganizationAccess was already the real (and only) org
  decision. Drop the unused param and the "org_membership" reason that
  getAccessAllowReason could never return.
- Replace boolean buildSignInDecision (re-evaluated for its log reason
  behind an unreachable `?? "allowlist"` fallback) with a single
  getStaticSignInReason that returns the reason once.
- Extract the GitHub org-membership HTTP client into
  github-org-membership.ts so access-control.ts is pure policy again.
- Collapse missing/unexpected membership-state logging into one branch;
  co-locate getOrgMembershipDecisionReason with the sign-in helpers.

No change to sign-in behavior. typecheck + web suite (470) green.
